requirements.txt
flask
werkzeug
torch
torchvision
pillow
requests
numpy
gunicorn
tqdm
datasets